requirements.txt
defusedxml~=0.7.1
ipython~=8.18.1
numpy~=2.0.2
pillow~=11.3.0
jupyterlab_pygments~=0.3.0
pip~=25.2
attrs~=25.3.0
pytz~=2025.2
Django~=4.2.25
asgiref~=3.10.0
sqlparse~=0.5.3
colorama~=0.4.6
PyYAML~=6.0.2
Jinja2~=3.1.6
requests~=2.32.5
djangorestframework~=3.16.1
Pygments~=2.19.2